Don't replicate entire directory structure from github if '.' is specified
If the added URL is a GH path, only keep the relative paths from that root url down, not the entire origin repo paths, if `.` is specified as the target path. The default behavior has been reverted to the previous one, where by default we'd preserve the entire upstream repo structure, which makes more sense, I think.

If the optional path after the URL is something other than `.`., it can still be used to change the destination root directory and simplify the folder hierarchy.