Simplify config by reusing `file` section directly also for glob URLs
We can fairly easily tell appart the glob url scenario from the regular one, and the way the behavior is currently implemented will nevertheless result in the same GH CLI probing anyway. So there was really no point in having a separate section for it.

The new design means that if you have a simple config like the following:

```
[file]
 url = https://github.com/dotnet/runtime/tree/master/src/libraries/Microsoft.Extensions.Configuration.Ini
 url = https://github.com/dotnet/runtime/tree/master/src/libraries/Microsoft.Extensions.Configuration.Json
 url = https://github.com/dotnet/runtime/tree/master/src/libraries/Microsoft.Extensions.Configuration.EnvironmentVariables
```

you can directly run `file update` and download all those three subdirectories. The file will be updated with the individually downloaded files and their etags as before, but since those URLs will remain in the file, the next update will automatically cause the GH probing, expanding and subsequent updating automatically, including adding new files :).

We might want to add a `file prune` for those scenarios, to keep the directory clean when files are deleted from the upstream URLs.