Workaround issue with conditional HTTP get for raw files
For some reason, raw URLs when retrieved with a GET + etag header, would return a new etag and a full response. Doing a HEAD + check etag in the response, with the *same* URL, however, works consistently :(.

So, we pay the price of two requests for each entry (HEAD first, GET second), but this avoids actually downloading a ton of unnecessary files. Seems like an acceptable trade-off anyway.