Fix sync operation which failed during GH CLI
We were simply returning true/false as success/failure from GH CLI, but a blob URL (meaning a 404 for a deleted upstream file) would also be considered false and cause a failure and stop processing.

By adding an enum and differentiating the cases more granularly, we can continue processing in that case.

Fixes #16