Revert behavior for entire repo processing
When we introduced the check for the lengh of the parts in the URI to only convert to raw when there was more than just org/repo, we broke the existing code in AddCommand that relied on a BadRequest response in that case to fallback to gh CLI processing.

This brings back that behavior with an explicit note so we don't break it again in the future.