Fix wrong sha for blob instead of tip commit
The API we were using (`contents`) returned the sha of the blob itself, instead of the tip commit for the file.

This fixes it so we keep the previous logic for mapping to the raw url, and just adds the commit if the raw URL request was successful (as an optimization).

Fixes #28.