Allow generating a changelog for updated/synced files
Adds a new `-c|changelog` argument where you can specify an output markdown file to write the changes to (defaults to `dotnet-file.md` if only `-c` is specified). The markdown is a full changelog built by comparing the old and new sha for changed files (if the sha is available) using the GH CLI (to avoid having to configure an access token), since that's the same mechanism we already use for GH fetching of directories.

We do some minor emoji cleanup too since the REST API does not properly support them for now and instead returns multiple chars for them. Might be worth revisting/removing down the road, but commits stick to latin subset, it should work just fine.

Fixes #30