Detect fetching folders/trees to avoid unnecessary auth
The authentication step is only needed when actually fetching raw github resources that potentially exist. But raw URLs pointing to `/tree/` will always fail by definition (can't get the raw content of a folder).

Since the raw handler overwrites the request URI, we preserve the original URI as the Referrer, so it can be checked by other handlers.